+++ adx/resource_adx_function.go
@@ -31,6 +31,9 @@ func resourceADXFunction() *schema.Resource {
 		UpdateContext: resourceADXFunctionUpdate,
 		ReadContext:   resourceADXFunctionRead,
 		DeleteContext: resourceADXFunctionDelete,
+		Importer: &schema.ResourceImporter{
+			StateContext: schema.ImportStatePassthroughContext,
+		},
 
 		Schema: map[string]*schema.Schema{
 			"cluster": getClusterConfigInputSchema(),